pyproject.toml
[tool.poetry]
name = "ghost_post_react"
version = "0.1.0"
description = ""
authors = ["Cedric Mulvihill <cedric.mulvihill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"

[tool.poetry.dev-dependencies]
djangorestframework = "^3.12.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
